test-requirements.txt
odoo_test_helper
odoo14-addon-product_state @git+https://github.com/OCA/product-attribute@refs/pull/931/head#subdirectory=setup/product_state
odoo14-addon-product_tier_validation @git+https://github.com/OCA/product-attribute@refs/pull/931/head#subdirectory=setup/product_tier_validation
